Refactor: Hide design settings (#7225)
