+++ src/DonationForms/FormDesigns/MultiStepFormDesign/css/_header.scss
@@ -11,157 +11,3 @@
         margin-bottom: var(--givewp-spacing-10);
     }
 }
-
-.givewp-design-settings--image-style {
-    &__background {
-        .givewp-layouts-header {
-            position: relative;
-            background: none;
-            padding-bottom: 25rem;
-            margin-bottom: 2.5rem;
-
-            &::before {
-                content: "";
-                position: absolute;
-                top: -2.5%;
-                left: -2rem;
-                width: calc(100% + 4rem);
-                height: 60%;
-                --tint-color: rgba(0, 0, 0, 0.3); /* Note: Color needs to be converted to RGB values. */
-                --tint: linear-gradient(90deg, var(--tint-color), var(--tint-color));
-                background-color: rgb(30, 140, 190);
-                background-image: var(--givewp-design-settings-background-image);
-                background-position: 50%;
-                background-size: cover;
-            }
-
-
-            .givewp-layouts-headerTitle h2, .givewp-layouts-headerDescription p {
-                position: absolute;
-                width: 100%;
-                height: 0;
-                color: var(--givewp-shades-white);
-                z-index: 999;
-            }
-
-            .givewp-layouts-headerTitle h2 {
-                top: 12.5%;
-            }
-
-            .givewp-layouts-headerDescription p {
-                top: 25%;
-            }
-        }
-
-        .givewp-layouts-goal {
-            position: absolute;
-            bottom: 0;
-        }
-
-        .givewp-donation-form__steps-body {
-            padding-top: 0;
-        }
-
-        .givewp-layouts-multiStepForm__form {
-            margin-top: 1.5rem;
-        }
-    }
-
-    &__cover {
-        /*
-         * Note: This requires that the Header.tsx template wrap the title/description.
-         *       The below examples adds the `.givewp-layouts-headerCover` wrapper.
-         */
-        &.givewp-donation-form {
-            position: relative;
-        }
-
-        .givewp-donation-form__steps {
-            overflow: visible;
-            background: var(--givewp-shades-white);
-            margin-top: 15rem;
-            box-shadow: none !important;
-
-            &-body {
-                border-radius: 0.5rem;
-                background: var(--givewp-shades-white);
-            }
-
-            &-footer-secure {
-                background: var(--givewp-shades-white);
-                border-radius: 0 0 0.5rem 0.5rem;
-            }
-        }
-
-        .givewp-layouts-headerTextWrapper {
-            position: absolute;
-            top: -50%;
-            left: 0;
-            width: 100%;
-
-            .givewp-layouts-headerTitle h2, .givewp-layouts-headerDescription p {
-                color: var(--givewp-shades-white);
-            }
-
-            :before {
-                content: "";
-                position: absolute;
-                top: -40%;
-                left: -23%;
-                width: 145%;
-                height: 24rem;
-                --tint-color: rgba(0, 0, 0, 0.3); /* Note: Color needs to be converted to RGB values. */
-                --tint: linear-gradient(90deg, var(--tint-color), var(--tint-color));
-                background-color: rgb(30, 140, 190);
-                background-image: var(--tint, linear-gradient(transparent, transparent)), var(--givewp-design-settings-background-image);
-                background-size: cover;
-                z-index: -100;
-            }
-
-            .givewp-layouts-headerTitle {
-                margin-bottom: var(--givewp-spacing-4);
-            }
-        }
-    }
-
-    &__above {
-        .givewp-donation-form__steps-body {
-            padding-top: 0;
-        }
-
-        .givewp-layouts-header > div:first-child {
-            padding-top: var(--givewp-spacing-6);
-        }
-
-        .givewp-layouts-header::before {
-            content: "";
-            position: absolute;
-            top: -320px;
-            left: -4rem;
-            min-width: calc(100% + 6rem);
-            width: 100%;
-            height: 320px;
-            background-image: var(--givewp-design-settings-background-image);
-            background-position: 50%;
-            background-size: cover;
-            border-top-left-radius: 0.5rem;
-            border-top-right-radius: 0.5rem;
-        }
-
-        .givewp-layouts-multiStepForm__form {
-            margin-top: 1.5rem;
-        }
-    }
-
-    &__center {
-        .givewp-donation-form__steps-body {
-            padding-top: 1.5rem;
-        }
-
-        .givewp-layouts-goal {
-            position: relative;
-            margin-top: calc(350px + var(--givewp-spacing-2));
-            border-radius: 0.5rem;
-        }
-    }
-}
