+++ src/DonationForms/ViewModels/DonationFormViewModel.php
@@ -104,8 +104,6 @@ class DonationFormViewModel
             ":root {
             --givewp-primary-color:{$this->primaryColor()};
             --givewp-secondary-color:{$this->secondaryColor()};
-            --givewp-design-settings-logo:url('{$this->formSettings->designSettingsLogoUrl}');
-            --givewp-design-settings-background-image:url('{$this->formSettings->designSettingsImageUrl}');
             }"
         );
 
@@ -188,7 +186,7 @@ class DonationFormViewModel
         $totalCountLabel = $this->getCountLabel($goalType);
 
         return [
-            'totalRevenue'    => $totalRevenue,
+            'totalRevenue' => $totalRevenue,
             'totalCountValue' => $totalCountValue,
             'totalCountLabel' => $totalCountLabel,
         ];
@@ -213,25 +211,25 @@ class DonationFormViewModel
         $formDesign = $this->getFormDesign($this->designId());
 
         return [
-            'donateUrl'            => $donateUrl,
-            'validateUrl'          => $validateUrl,
-            'authUrl'              => $authUrl,
+            'donateUrl' => $donateUrl,
+            'validateUrl' => $validateUrl,
+            'authUrl' => $authUrl,
             'inlineRedirectRoutes' => [
                 'donation-confirmation-receipt-view',
             ],
-            'registeredGateways'   => $formDataGateways,
-            'form'                 => array_merge($formApi->jsonSerialize(), [
+            'registeredGateways' => $formDataGateways,
+            'form' => array_merge($formApi->jsonSerialize(), [
                 'settings' => $this->formSettings,
                 'currency' => $formApi->getDefaultCurrency(),
-                'goal'     => $donationFormGoalData->toArray(),
-                'stats'    => $this->formStatsData(),
-                'design'   => $formDesign ? [
-                    'id'          => $formDesign::id(),
-                    'name'        => $formDesign::name(),
+                'goal' => $donationFormGoalData->toArray(),
+                'stats' => $this->formStatsData(),
+                'design' => $formDesign ? [
+                    'id' => $formDesign::id(),
+                    'name' => $formDesign::name(),
                     'isMultiStep' => $formDesign->isMultiStep(),
                 ] : null,
             ]),
-            'previewMode'          => $this->previewMode,
+            'previewMode' => $this->previewMode,
         ];
     }
 
@@ -276,21 +274,6 @@ class DonationFormViewModel
         if ($this->previewMode) {
             $classNames[] = 'givewp-donation-form--preview';
         }
-
-        if ($this->formSettings->designSettingsImageUrl) {
-            $classNames[] = 'givewp-design-settings--image';
-            $classNames[] = 'givewp-design-settings--image-style__' . $this->formSettings->designSettingsImageStyle;
-        }
-
-        if ($this->formSettings->designSettingsLogoUrl) {
-            $classNames[] = 'givewp-design-settings--logo';
-            $classNames[] = 'givewp-design-settings--logo-position__' . $this->formSettings->designSettingsLogoPosition;
-        }
-
-        $classNames[] = 'givewp-design-settings--section-style__' . $this->formSettings->designSettingsSectionStyle;
-
-        $classNames[] = 'givewp-design-settings--textField-style__' . $this->formSettings->designSettingsTextFieldStyle;
-
         ?>
 
         <div data-theme="light" id="root-givewp-donation-form"
@@ -447,5 +430,25 @@ class DonationFormViewModel
         );
 
         Language::setScriptTranslations($handle);
+    }
+
+    /**
+     * @unreleased
+     */
+    private function updateDesignSettingsClassNames(array &$classNames)
+    {
+        if ($this->formSettings->designSettingsImageUrl) {
+            $classNames[] = 'givewp-design-settings--image';
+            $classNames[] = 'givewp-design-settings--image-style__' . $this->formSettings->designSettingsImageStyle;
+        }
+
+        if ($this->formSettings->designSettingsLogoUrl) {
+            $classNames[] = 'givewp-design-settings--logo';
+            $classNames[] = 'givewp-design-settings--logo-position__' . $this->formSettings->designSettingsLogoPosition;
+        }
+
+        $classNames[] = 'givewp-design-settings--section-style__' . $this->formSettings->designSettingsSectionStyle;
+
+        $classNames[] = 'givewp-design-settings--textField-style__' . $this->formSettings->designSettingsTextFieldStyle;
     }
 }
