+++ src/DonationForms/resources/styles/base.scss
@@ -29,8 +29,3 @@
 @layer base-overrides {
     @import "base-overrides";
 }
-
-@import "design-settings/image";
-@import "design-settings/logo";
-@import "design-settings/section";
-@import "design-settings/text-field";
