+++ src/FormBuilder/resources/js/form-builder/src/settings/design/general-controls/header/index.tsx
@@ -1,8 +1,6 @@
 import {__} from '@wordpress/i18n';
-import {PanelBody, PanelRow, SelectControl, TextareaControl, TextControl, ToggleControl} from '@wordpress/components';
+import {PanelBody, PanelRow, TextareaControl, TextControl, ToggleControl} from '@wordpress/components';
 import {setFormSettings, useFormState} from '@givewp/form-builder/stores/form-state';
-import {upload} from '@wordpress/icons';
-import MediaLibrary from '@givewp/form-builder/components/settings/MediaLibrary';
 
 export default function Header({dispatch, publishSettings}) {
     const {
@@ -12,24 +10,8 @@ export default function Header({dispatch, publishSettings}) {
             heading,
             showDescription,
             description,
-            designSettingsImageUrl,
-            designSettingsImageStyle,
-            designSettingsLogoUrl,
-            designSettingsLogoPosition,
         },
     } = useFormState();
-
-    const resetSettings = () => {
-        const reset = {
-            designSettingsImageUrl: '',
-            designSettingsImageStyle: '',
-            designSettingsLogoUrl,
-            designSettingsLogoPosition,
-        };
-
-        dispatch(setFormSettings(reset));
-        publishSettings(reset);
-    };
 
     return (
         <PanelBody title={__('Header', 'give')} initialOpen={false}>
@@ -40,19 +22,11 @@ export default function Header({dispatch, publishSettings}) {
                     onChange={() => {
                         dispatch(
                             setFormSettings({
-                                showHeader: !showHeader,
-                                designSettingsImageUrl,
-                                designSettingsImageStyle,
-                                designSettingsLogoUrl,
-                                designSettingsLogoPosition,
+                                showHeader: !showHeader
                             })
                         );
                         publishSettings({
-                            showHeader: !showHeader,
-                            designSettingsImageUrl,
-                            designSettingsImageStyle,
-                            designSettingsLogoUrl,
-                            designSettingsLogoPosition,
+                            showHeader: !showHeader
                         });
                     }}
                 />
@@ -67,19 +41,11 @@ export default function Header({dispatch, publishSettings}) {
                             onChange={() => {
                                 dispatch(
                                     setFormSettings({
-                                        showHeading: !showHeading,
-                                        designSettingsImageUrl,
-                                        designSettingsImageStyle,
-                                        designSettingsLogoUrl,
-                                        designSettingsLogoPosition,
+                                        showHeading: !showHeading
                                     })
                                 );
                                 publishSettings({
-                                    showHeading: !showHeading,
-                                    designSettingsImageUrl,
-                                    designSettingsImageStyle,
-                                    designSettingsLogoUrl,
-                                    designSettingsLogoPosition,
+                                    showHeading: !showHeading
                                 });
                             }}
                         />
@@ -91,19 +57,11 @@ export default function Header({dispatch, publishSettings}) {
                             onChange={() => {
                                 dispatch(
                                     setFormSettings({
-                                        showDescription: !showDescription,
-                                        designSettingsImageUrl,
-                                        designSettingsImageStyle,
-                                        designSettingsLogoUrl,
-                                        designSettingsLogoPosition,
+                                        showDescription: !showDescription
                                     })
                                 );
                                 publishSettings({
-                                    showDescription: !showDescription,
-                                    designSettingsImageUrl,
-                                    designSettingsImageStyle,
-                                    designSettingsLogoUrl,
-                                    designSettingsLogoPosition,
+                                    showDescription: !showDescription
                                 });
                             }}
                         />
@@ -116,19 +74,11 @@ export default function Header({dispatch, publishSettings}) {
                                 onChange={(heading) => {
                                     dispatch(
                                         setFormSettings({
-                                            heading,
-                                            designSettingsImageUrl,
-                                            designSettingsImageStyle,
-                                            designSettingsLogoUrl,
-                                            designSettingsLogoPosition,
+                                            heading
                                         })
                                     );
                                     publishSettings({
-                                        heading,
-                                        designSettingsImageUrl,
-                                        designSettingsImageStyle,
-                                        designSettingsLogoUrl,
-                                        designSettingsLogoPosition,
+                                        heading
                                     });
                                 }}
                             />
@@ -142,82 +92,13 @@ export default function Header({dispatch, publishSettings}) {
                                 onChange={(description) => {
                                     dispatch(
                                         setFormSettings({
-                                            description,
-                                            designSettingsImageUrl,
-                                            designSettingsImageStyle,
-                                            designSettingsLogoUrl,
-                                            designSettingsLogoPosition,
+                                            description
                                         })
                                     );
                                     publishSettings({
-                                        description,
-                                        designSettingsImageUrl,
-                                        designSettingsImageStyle,
-                                        designSettingsLogoUrl,
-                                        designSettingsLogoPosition,
+                                        description
                                     });
                                 }}
-                            />
-                        </PanelRow>
-                    )}
-                    <PanelRow>
-                        <MediaLibrary
-                            id="givewp-header-media-library-control"
-                            icon={upload}
-                            label={__('Image', 'give')}
-                            actionLabel={__('Upload Image', 'give')}
-                            value={designSettingsImageUrl}
-                            onChange={(designSettingsImageUrl) => {
-                                dispatch(
-                                    setFormSettings({
-                                        designSettingsImageUrl,
-                                        designSettingsImageStyle,
-                                        designSettingsLogoUrl,
-                                        designSettingsLogoPosition,
-                                    })
-                                );
-
-                                publishSettings({
-                                    designSettingsImageUrl,
-                                    designSettingsImageStyle,
-                                    designSettingsLogoUrl,
-                                    designSettingsLogoPosition,
-                                });
-                            }}
-                            reset={resetSettings}
-                        />
-                    </PanelRow>
-                    {designSettingsImageUrl && (
-                        <PanelRow>
-                            <SelectControl
-                                label={__('Image Style', 'give')}
-                                help={__(
-                                    'Determines how the image will be applied in the header for this form.',
-                                    'give'
-                                )}
-                                onChange={(designSettingsImageStyle) => {
-                                    dispatch(
-                                        setFormSettings({
-                                            designSettingsImageUrl,
-                                            designSettingsImageStyle,
-                                            designSettingsLogoUrl,
-                                            designSettingsLogoPosition,
-                                        })
-                                    );
-
-                                    publishSettings({
-                                        designSettingsImageUrl,
-                                        designSettingsImageStyle,
-                                        designSettingsLogoUrl,
-                                        designSettingsLogoPosition,
-                                    });
-                                }}
-                                value={designSettingsImageStyle}
-                                options={[
-                                    {label: __('Background', 'give'), value: 'background'},
-                                    {label: __('Above', 'give'), value: 'above'},
-                                    {label: __('Center', 'give'), value: 'center'},
-                                ]}
                             />
                         </PanelRow>
                     )}
